requirements.txt
Flask
Flask-SQLAlchemy
python-dotenv
pytest
gunicorn
flasgger
psycopg2-binary
pdfplumber
pytesseract
opencv-python
pyzbar
Pillow
PyJWT
flask-cors